PII join api changes
