--- generics/constants/enum.utils.js
@@ -1,16 +1,0 @@
-/**
- * name : enum.utils.js
- * author : Ankit Shahu
- * Date : 24-jan-2023
- * Description : Enums are stored in this file.
- */
-
-
-exports.Consent = {
-    ACTIVE:"ACTIVE",
-    REVOKED:"REVOKED"
-}
-
-exports.objectType = {
-    PROGRAM:"Program"
-}
