--- generics/constants/schema.js
@@ -1,15 +1,0 @@
-/**
- * name : constants/schema.js
- * author : Aman
- * Date : 28-feb-2020
- * Description : All common schema information.
- */
-
-module.exports = {
-    METAINFORMATION : "metaInformation",
-    EXTERNALID : "externalId",
-    NAME : "name",
-    ENTITYTYPE : "entityType",
-    GROUPS : "groups",
-    ID : "_id"
-}
