+++ generics/services/users.js
@@ -345,11 +345,70 @@ async function getParentEntities( entityId, iteration = 0, parentEntities ) {
     return parentEntities;
 
 }
+
+/**
+  * update user consent for sharing the PII.
+  * @method
+  * @name consent
+  * @param {String} token - user token
+  * @returns {Object} consentData - consent data.
+*/
+
+const consent = function ( token, consentData ) {
+    return new Promise(async (resolve, reject) => {
+        try {
+            
+            let url = userServiceUrl + constants.endpoints.CONSENT_API;
+
+            const options = {
+                headers : {
+                    "content-type": "application/json",
+                    "x-authenticated-user-token" : token,
+                    "Authorization" : process.env.SUNBIRD_SERVICE_AUTHERIZATION
+                },
+                body: JSON.stringify(consentData) 
+            };
+            
+            request.post(url,options,kendraCallback);
+
+            function kendraCallback(err, data) {
+
+                let result = {
+                    success : true
+                };
+
+                if (err) {
+                    result.success = false;
+                } else {
+                    
+                    let response = JSON.parse(data.body);
+                    if( response.responseCode === httpStatusCode['http_responsecode_ok'].message ) {
+                        result["data"] = response;
+                    } else {
+                        result["message"] = response;
+                        result.success = false;
+                    }
+
+                }
+                return resolve(result);
+            }
+            setTimeout(function () {
+                return resolve (result = {
+                    success : false
+                });
+            }, constants.common.SERVER_TIME_OUT);
+
+        } catch (error) {
+            return reject(error);
+        }
+    })
+}
   
 module.exports = {
     profile : profile,
     locationSearch : locationSearch,
     orgSchoolSearch :orgSchoolSearch,
     getSubEntitiesBasedOnEntityType : getSubEntitiesBasedOnEntityType,
-    getParentEntities : getParentEntities
+    getParentEntities : getParentEntities,
+    consent : consent
 }
